+++ src/plugins/index.ts
@@ -4,6 +4,10 @@ import * as wikipedia from './wikipedia.js';
 import * as branchIoDeeplinks from './branchio-deeplinks.js';
 import * as youtube from './youtube.js';
 import * as spotify from './spotify.js';
+import * as dlsite from './dlsite.js';
+import * as iwara from './iwara.js';
+import * as komiflo from './komiflo.js';
+import * as nijie from './nijie.js';
 import { SummalyPlugin } from '@/iplugin.js';
 
 export const plugins: SummalyPlugin[] = [
@@ -13,4 +17,8 @@ export const plugins: SummalyPlugin[] = [
 	branchIoDeeplinks,
 	youtube,
 	spotify,
+	dlsite,
+	iwara,
+	komiflo,
+	nijie,
 ];